requirements.txt
python-ldap >= 2.4.19
